requirements.txt
autopep8==1.5.4
bcrypt==3.1.7
blinker==1.4
cffi==1.14.1
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
flake8==3.8.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.1
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
Werkzeug==1.0.1
WTForms==2.3.3
